Add support for waiting before starting jobs.
Implement a mechanism (for reasons of backwards compatibility)
to wait some time before starting a job. Waiting is done per host,
ie. it is guaranteed that inbetween consecutive jobs assigned to
a specific host a certain number of seconds is waited.

The number of seconds is configurable using the `--setup-pause`
cmdline argument. This option defaults to 0.